# Conflicts:
#	scripts/globals/bcnm.lua
#	scripts/globals/mobskills/eald2_warp_in.lua
#	scripts/globals/mobskills/eald2_warp_out.lua
#	scripts/globals/mobskills/earth_blade.lua
#	scripts/globals/mobskills/fire_blade.lua
#	scripts/globals/mobskills/frost_blade.lua
#	scripts/globals/mobskills/lightning_blade.lua
#	scripts/globals/mobskills/water_blade.lua
#	scripts/globals/mobskills/wind_blade2.lua
#	scripts/zones/Empyreal_Paradox/npcs/Transcendental_Radiance.lua
#	scripts/zones/Grand_Palace_of_HuXzoi/npcs/_iya.lua
#	scripts/zones/Lower_Jeuno/npcs/Aldo.lua
#	scripts/zones/Norg/npcs/Gilgamesh.lua
#	scripts/zones/Norg/npcs/_700.lua
#	scripts/zones/RuLude_Gardens/IDs.lua
#	scripts/zones/Sealions_Den/Zone.lua
#	sql/mob_groups.sql
#	sql/mob_pools.sql
#	sql/mob_skill_lists.sql
#	sql/mob_spell_lists.sql

+++ src/map/merit.cpp
@@ -462,7 +462,7 @@ void CMeritPoints::LowerMerit(MERIT_TYPE merit)
 int32 CMeritPoints::GetMeritValue(MERIT_TYPE merit, CCharEntity* PChar)
 {
     Merit_t* PMerit = GetMeritPointer(merit);
-	uint8 meritValue = 0;
+	uint16 meritValue = 0;
 
     if (PMerit)
     {
